refactor: extract shared buildPivotFilters from duplicated De Morgan logic
Eliminates ~40 lines of duplicated tuple-exclusion filter logic between
drillDownFilters.ts and UnderlyingDataModal.tsx by extracting a shared
buildPivotFilters utility. Both consumers now use the same tested
implementation for Other group drill-through filters.
