+++ packages/frontend/src/components/MetricQueryData/drillDownFilters.ts
@@ -9,6 +9,7 @@ import {
     type ResultValue,
 } from '@lightdash/common';
 import { v4 as uuidv4 } from 'uuid';
+import { buildPivotFilters } from './pivotFilters';
 import type { TopGroupTuple } from './types';
 
 type CombineFiltersArgs = {
@@ -36,69 +37,10 @@ export const combineFilters = ({
     if (dashboardFilters) {
         combinedDimensionFilters.push(...dashboardFilters.dimensions);
     }
-    if (pivotReference?.pivotValues) {
-        const hasOtherPivot = pivotReference.pivotValues.some(
-            (pv) => pv.isOtherGroup,
+    if (pivotReference) {
+        combinedDimensionFilters.push(
+            ...buildPivotFilters({ pivotReference, topGroupTuples }),
         );
-
-        if (hasOtherPivot && topGroupTuples?.length) {
-            // Build tuple-aware exclusion: NOT((A₁ AND B₁) OR (A₂ AND B₂))
-            // Via De Morgan: (NOT A₁ OR NOT B₁) AND (NOT A₂ OR NOT B₂)
-            const tupleExclusions: FilterGroupItem[] = topGroupTuples.map(
-                (tuple) => ({
-                    id: uuidv4(),
-                    or: Object.entries(tuple).map(
-                        ([field, tupleValue]): FilterRule => ({
-                            id: uuidv4(),
-                            target: { fieldId: field },
-                            operator:
-                                tupleValue === null
-                                    ? FilterOperator.NOT_NULL
-                                    : FilterOperator.NOT_EQUALS,
-                            values:
-                                tupleValue === null
-                                    ? undefined
-                                    : [tupleValue as string],
-                        }),
-                    ),
-                }),
-            );
-            combinedDimensionFilters.push(...tupleExclusions);
-
-            // Also add non-Other pivot values as equality filters
-            const nonOtherPivots = pivotReference.pivotValues.filter(
-                (pv) => !pv.isOtherGroup,
-            );
-            combinedDimensionFilters.push(
-                ...nonOtherPivots.map(
-                    (pivot): FilterRule => ({
-                        id: uuidv4(),
-                        target: { fieldId: pivot.field },
-                        operator:
-                            pivot.value === null
-                                ? FilterOperator.NULL
-                                : FilterOperator.EQUALS,
-                        values:
-                            pivot.value === null ? undefined : [pivot.value],
-                    }),
-                ),
-            );
-        } else {
-            const pivotFilter: FilterRule[] = pivotReference.pivotValues
-                .filter((pv) => !pv.isOtherGroup)
-                .map((pivot) => ({
-                    id: uuidv4(),
-                    target: {
-                        fieldId: pivot.field,
-                    },
-                    operator:
-                        pivot.value === null
-                            ? FilterOperator.NULL
-                            : FilterOperator.EQUALS,
-                    values: pivot.value === null ? undefined : [pivot.value],
-                }));
-            combinedDimensionFilters.push(...pivotFilter);
-        }
     }
     if (extraFilters?.dimensions) {
         combinedDimensionFilters.push(extraFilters.dimensions);
